Added fieldAlias option
When using the OnetoMany relation we often do not want the parameter name as the fieldname in the Solr database. For example..

Company -> Orders ($orders) -> getOrderNumber()

Here the fieldname will be casted to orders_ss in the solr_database. Instead we want order_id_ss.

To arrange this behaviour we added the fieldAlias in the annotation.

 * @Solr\Field(type="strings", getter="getOrderId", fieldAlias="order_id")