Dear Florian,
Thank you for checking my changes. The problem lies in the fact that the current logic for the EntityIndexerSubsriber binds to the postPersist event. I have changes this to the onFlush event so we have access to all the relational data. Now the id's get persisted as expected.

Changes per file:

AnnotationReader: Allow empty relations to be nullable

DocumentFactory: FieldValue should not be required

EntityIndexerSubscriber: Changed the indexer to onFlush instead of postPersist. At that point the ORM relations are available.

Can you please (re)reevaluate my PR? If you require any further information, feel free to contact me.

Kind regards,

Jack Kwakman